Always show nav bar, override requestURL, added static web VC, show links in modal
