+++ Wikivoyage/LocationWebViewController.swift
@@ -6,46 +6,34 @@
 //  Copyright (c) 2015 Ben Meline. All rights reserved.
 //
 
-import UIKit
-import WebKit
 import Alamofire
 import SwiftyJSON
 import MagicalRecord
 
-class LocationWebViewController: WebViewController {
-
-    var originalURLSet: Bool = false
-    var originalHost: String?
-    var originalPath: String?
-    @IBOutlet var saveButton: UIBarButtonItem!
+class LocationWebViewController: StaticWebViewController {
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        saveButton.enabled = false
+    @IBAction func favorite(sender: AnyObject) {
+        favoritePage()
     }
     
-    override func webView(webView: WKWebView, didCommitNavigation navigation: WKNavigation!) {
-        super.webView(webView, didCommitNavigation: navigation)
-        
-        if !originalURLSet {
-            originalHost = webView.URL?.host
-            originalPath = webView.URL?.path
-            originalURLSet = true
-        }
-
-        if webView.URL?.host == originalHost && webView.URL?.path == originalPath {
-            saveButton.enabled = true
+    @IBAction func download(sender: AnyObject) {
+        downloadPage()
+    }
+    
+    func favoritePage() {
+        let id = NSNumber(integer: self.pageId)
+        if let savedPage = SavedPage.MR_findFirstByAttribute("id", withValue: id) {
+            // It exists
         } else {
-            saveButton.enabled = false
+            let savedPage = SavedPage.MR_createEntity()
+            savedPage.title = self.pageTitle
+            savedPage.id = self.pageId
         }
+        
+        NSManagedObjectContext.MR_defaultContext().MR_saveToPersistentStoreAndWait()
     }
     
-    // Save page
-    @IBAction func savePage(sender: AnyObject) {
-        downloadText()
-    }
-    
-    func downloadText() {
+    func downloadPage() {
         Alamofire.request(.GET, "http://en.wikivoyage.org/w/api.php", parameters: ["action": "parse", "pageid": pageId, "prop": "text", "mobileformat": "", "noimages": "", "format": "json"]).responseJSON() {
             (_, _, data, error) in
             if(error != nil) {
@@ -53,10 +41,9 @@ class LocationWebViewController: WebViewController {
             } else {
                 let json = JSON(data!)
                 
-                // Check if it's already saved and if currently displaying Wikivoyage page
+                // Check if it's already saved and save HTML
                 let id = NSNumber(integer: self.pageId)
                 if let savedPage = SavedPage.MR_findFirstByAttribute("id", withValue: id) {
-                    savedPage.title = self.pageTitle
                     savedPage.html = json["parse"]["text"]["*"].stringValue
                 } else {
                     let savedPage = SavedPage.MR_createEntity()
@@ -69,22 +56,4 @@ class LocationWebViewController: WebViewController {
             }
         }
     }
-    
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
