--- Wikivoyage/Wikivoyage-Bridging-Header.h
@@ -1,12 +1,0 @@
-//
-//  Wikivoyage-Bridging-Header.h
-//  Wikivoyage
-//
-//  Created by Ben Meline on 9/1/15.
-//  Copyright (c) 2015 Ben Meline. All rights reserved.
-//
-
-#ifndef Wikivoyage_Wikivoyage_Bridging_Header_h
-#define Wikivoyage_Wikivoyage_Bridging_Header_h
-
-#endif
